backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
opencv-python==4.5.3.56
requests==2.26.0
roboflow==0.2.29
numpy==1.21.2
Pillow==8.3.2 